requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
twilio==8.10.0
python-dotenv==1.0.0
